Disable ELF check for libimsma
The ELF check reports that a "libsink" dependency is missing for the armv-9a architecture that the libimsma.so library requires according to its ELF dependency table (DT_NEEDED). However the precompiled library can work without it and disabling the check prevents unnecessary errors and allows the compilation to continue.